requirements.txt
certifi==2018.8.24
chardet==3.0.4
click==6.7
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
idna==2.7
lxml==4.2.4
Markdown==3.0.1
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==4.1
